moving repo favorite api calls to reusable hook
